Everything works except the interrupt endpoints don't give interrupts

+++ kernel/src/x86/apic.h
@@ -7,8 +7,8 @@
 
 namespace Apic {
   void init();
-  void write(uint8_t reg, uint32_t value);
-  uint32_t read(uint8_t reg);
+  void write(uint8_t reg, uint64_t value);
+  uint64_t read(uint8_t reg);
   void write_ICR(uint64_t value);
   uint64_t read_ICR();
 }
